backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.10.0
pydantic-settings>=2.4.0
email-validator>=2.0.0

# Database
supabase>=2.3.0

# Queue/Workers
redis>=5.0.0
rq>=1.15.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
python-multipart>=0.0.6

# HTTP Client
aiohttp>=3.9.0
httpx>=0.26.0

# Payments
stripe>=7.0.0

# AI/ML Services
google-generativeai>=0.8.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
hypothesis>=6.92.0
pytest-cov>=4.1.0

# Linting/Formatting
ruff>=0.1.0
black>=23.12.0
mypy>=1.8.0
